Move getEmptyMockSession() helper to a helper file that storybook can also load.
