+++ app/stories/pages/how-it-works.stories.tsx
@@ -1,10 +1,10 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react'
-import cloneDeep from 'lodash/cloneDeep'
 
 import HowItWorksPage from '@pages/how-it-works'
 
 import { getBackRoute } from '@utils/routing'
-import { initialSessionData } from '@utils/sessionData'
+
+import { getEmptyMockSession } from '../../tests/helpers/mockData'
 
 export default {
   title: 'Pages/How It Works',
@@ -12,7 +12,7 @@ export default {
 } as ComponentMeta<typeof HowItWorksPage>
 
 const Template: ComponentStory<typeof HowItWorksPage> = (args) => {
-  args.backRoute = getBackRoute('/how-it-works', cloneDeep(initialSessionData))
+  args.backRoute = getBackRoute('/how-it-works', getEmptyMockSession())
   return <HowItWorksPage {...args} />
 }
