+++ app/tests/utils/routing/getForwardRoute.test.ts
@@ -5,7 +5,7 @@ import {
   getMockEligibilityData,
   getMockIncomeData,
 } from '../../helpers/mockData'
-import { getEmptyMockSession } from '../../helpers/setup'
+import { getEmptyMockSession } from '../../helpers/mockData'
 
 const simplePaths = [
   ['/', '/how-it-works'],
